Rework to use int type

+++ mudock/src/grid/grid_desolvation_map.cpp
@@ -22,9 +22,9 @@ namespace mudock {
     const point3D grid_maximum{std::ceil((receptor_max_x + cutoff_distance) * fp_type{2}) / fp_type{2},
                                std::ceil((receptor_max_y + cutoff_distance) * fp_type{2}) / fp_type{2},
                                std::ceil((receptor_max_z + cutoff_distance) * fp_type{2}) / fp_type{2}};
-    const index3D npts{static_cast<size_t>((grid_maximum.x - grid_minimum.x) / grid_spacing),
-                       static_cast<size_t>((grid_maximum.y - grid_minimum.y) / grid_spacing),
-                       static_cast<size_t>((grid_maximum.z - grid_minimum.z) / grid_spacing)};
+    const index3D npts{static_cast<int>((grid_maximum.x - grid_minimum.x) / grid_spacing),
+                       static_cast<int>((grid_maximum.y - grid_minimum.y) / grid_spacing),
+                       static_cast<int>((grid_maximum.z - grid_minimum.z) / grid_spacing)};
 
     grid_map desolvation_map{npts, grid_minimum, grid_maximum};
 
@@ -33,28 +33,28 @@ namespace mudock {
     * and will not be smoothed 
     */
     std::array<fp_type, NDIEL> sol_fn;
-    for (size_t indx_r = 1; indx_r < NDIEL; indx_r++) {
+    for (int indx_r = 1; indx_r < NDIEL; indx_r++) {
       const fp_type r = indx_r / A_DIV;
       sol_fn[indx_r] =
           autodock_parameters::coeff_desolv * std::exp(-(r * r) / (fp_type{2} * (sigma * sigma)));
     }
 
-    for (size_t index_z = 0; index_z < npts.size_z(); ++index_z) {
+    for (int index_z = 0; index_z < npts.size_z(); ++index_z) {
       /*
       *  c[0:2] contains the current grid point.
       */
       const fp_type coord_z = grid_minimum.z + index_z * grid_spacing;
-      for (size_t index_y = 0; index_y < npts.size_y(); ++index_y) {
+      for (int index_y = 0; index_y < npts.size_y(); ++index_y) {
         const fp_type coord_y = grid_minimum.y + index_y * grid_spacing;
-        for (size_t index_x = 0; index_x < npts.size_x(); ++index_x) {
+        for (int index_x = 0; index_x < npts.size_x(); ++index_x) {
           const fp_type coord_x = grid_minimum.x + index_x * grid_spacing;
           fp_type energy{0};
-          for (size_t index = 0; index < receptor.num_atoms(); ++index) {
+          for (int index = 0; index < receptor.num_atoms(); ++index) {
             const fp_type d = distance(point3D{receptor.x(index), receptor.y(index), receptor.z(index)},
                                        point3D{coord_x, coord_y, coord_z});
             if (d > NBC)
               continue; /* onto the next atom... */
-            const size_t indx_r = std::min<size_t>(std::floor(d * A_DIV), NDIEL - 1);
+            const int indx_r = std::min<int>(std::floor(d * A_DIV), NDIEL - 1);
             energy += solpar_q * receptor.get_vol()[index] * sol_fn[indx_r];
           }
           desolvation_map.at(index_x, index_y, index_z) = energy;
